Use latest platform versions by default
Remove all the default.json files and thus default to the newest (well,
based on filename sorting) version if `:version` option is not
specified. If the filename sorting is not valid with some platform in
the future, a default.json can be then added for that platform.

The whole feature is undocumented and users are encouraged to always
declare the specific version anyway.