Make chef/ohai a development dep
Fauxhai 3 can't be included in chefspec at the moment due to the Chef 12
dep.  Doing so would require us to remove Chef 11 support from
Chefspec. Chef 11 is still supported so that's not an acceptable option.
Since Ohai / Chef are only needed for dumping new platforms we should
move those to dev deps.