Generate a markdown list of platforms/versions via Rake
It comes up a lot in issues / slack that people don’t know what platforms are available to them. Yes it’s in the source, but this puts it up front in a really simple format. This needs to get wired into Travis so the list has to get updated before PRs go green. That’s not hard, but this is a super simple win.

Signed-off-by: Tim Smith <tsmith@chef.io>